Foodwise/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
clarifai-grpc==7.7.0
click==8.0.1
Flask==2.0.1
googleapis-common-protos==1.53.0
grpcio==1.39.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
protobuf==3.17.3
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
yake==0.4.8